Add files via upload
Changed a couple CSS ID's to classes as ID's can only be used once per page.  Classes can be used many times.  I also added a couple other classes or tags to handle CSS instead of putting it in the HTML

Also added a fineprint class to the login.css.